Move upgrade functions to separate file
